+++ src/Tabs/index.tsx
@@ -1,0 +1,4 @@
+import Tabs from './Tabs';
+
+export type { TabsProps } from './Tabs';
+export default Tabs;